Move rfc3339 tests to the `views` module
